Add test cases
