+++ src/Sulu/Bundle/MediaBundle/DependencyInjection/SuluMediaExtension.php
@@ -16,7 +16,6 @@ use FFMpeg\FFMpeg;
 use Imagine\Vips\Imagine as VipsImagine;
 use Sulu\Bundle\MediaBundle\Admin\MediaAdmin;
 use Sulu\Bundle\MediaBundle\Entity\Collection;
-use Sulu\Bundle\MediaBundle\Entity\CollectionInterface;
 use Sulu\Bundle\MediaBundle\Entity\MediaRepositoryInterface;
 use Sulu\Bundle\MediaBundle\Media\Exception\FileVersionNotFoundException;
 use Sulu\Bundle\MediaBundle\Media\Exception\FormatNotFoundException;
